Add a way to avoid reading the existing schema before each migration
Most migrations don't actually perform changes on the Schema object but
rather add SQL queries directly in the migration, because it is much
more powerful (it is the only way to migrate data for instance, as this
must generally happen in the middle of schema changes) and because the
diff commands generates migrations this way.
For such migrations, the Schema object will never be altered, but the
migration tool will still need to read it at the beginning of each
migration to pass it to the Migration (which does not care about it). To
improve performance, migrations can now implement an optional marker
interface. In such cases, the migration will receive an empty Schema
object and will not be allowed to perform changes on it. This avoids the
need to read the database schema all the time for nothing.