Fix issue of false warning for no queries after migrations
Migrations shows a warning if not actual SQL statement is run after
migrations. After merging "add-sql" branch
(Ydbbcee1906524f4667c9db06ee9afef476fa0)
an issue is introduced for the warning kept showing.

This is because the logic of executing registered SQL statements was
moved to a private function which is called once after the
migration and once after post migration method. This logic
(`executeRegisteredSql` mehtod) used to show the warning for no SQL
statements, and when called after post migration methods it is
quite possible to show the warnings, since they might not register
any SQL statements at all.

Fixed this issue by removing showing the warning from the
`executeRegisteredSql` method to the caller (`execute` method).

Also the previous behavior of `execute` method was to return the list of
SQL statements registered, but "add-sql" branch affected this, since
we cleared out the `sql` property of the `Version` class after executing
them. Fixed this issue by keeping a local copy of the SQL statements
and return that list instead.