Don't Prompt the User for Confirmation on the `migrations:migrate` Command When There is Nothing to Do (#480)
* Improve the Tests for the Migrate Command

This brings them inline with some of the other, newer command tests that
use Symfony's command tester. There's more coverage as well. Done so
some changes can be made to `MigrateCommand`.

* Bail out of MigrationCommand::execute Early on --write-sql

Removes a level of nesting from the rest of the method so some changes
can be a bit nicer later on.

* Move the "No Migrations..." Message into Migrations::migrate

Provides the same output to the user as before.

* Add a Confirmation Callback to Migration::migrate

Provides a way to hook into the migration and ask the user for
confirmation. Right now that happens further up stream in a console
command.

* Convert MigrateCommand to use the Confirmation Callback

Means there's no longer a big warning or confirmation when no migrations
are to be executed.

* Re-Add Some Newlines Between @param, @return, and @throws

* Don't Return `false` from Migrations::migrate on Cancel

Because it's not really an error.

This also reworks the console command to keep the same interface:
returns a 1 status code on cancel and outputs "migrations cancelled"