handle unregistered migrations in up-to-date check
we need to handle the case where there are extra migrations in doctrine. without this change, the command outputs something like "Out-of-date! 18446744073709551613 migration is available to execute." the > 1 check says false and we convert the signed integer value to unsigned.
with this fix, we explicitly handle the case. the default is to report the issue but not fail, same as the migrate command that only outputs a warning but does not refuse to run. i also add a flag in case people want a strict check.