Handle deleted migration classes #795
If a *Version.php class is deleted and that version has already been migrated and exists in the database,
the migrations:version --delete <version> command will fail because the version class does not exist.

To fix the failing command, the user will see and extended information message, gets the opportunity
to create an empty migration class and is then able to delete the specific version.