Depend directly on the symfony stopwatch component
Before this commit doctrine/migrations was using a wrapper on the symfony stopwatch component, but that wrapper was still returning the original symfony stopwatch event objects, making the wrapping almost useless as the doctrine code was still relying on the doctrine implementation. This commit just removes the wrapper and depends directly on the symfony stopwatch component.