Migrate exits with 0 for empty plans as in 2.x
In doctrine 2.x the migration command throws an exception if there
are not migrations to execute (unless the --allow-no-migrations
parameter is passed). In 3.0.0 this was misunderstood and the exception
was thrown even if there are registered migrations but there are no migrations to
execute for the current migration plan.
With this commit we are back to the 2.x strategy.

This commit also improves the error messages when the various aliases
can not be reached or are already reached.