Address MasterSlaveConnection rename
MasterSlaveConnection has been deprecated in favor of
PrimaryReadReplicaConnection, which it extends before being removed in
DBAL 3. This ensure both cases are handled when it comes to making sure
we connect to the primary, while not bumping the requirement to
doctrine/dbal 2.11 (which requires PHP 7.3, although subsequent version
might restore compatibility with 7.1).