Apply workaround to maintain expected behavior
This test case addition simulates what happens when the
`--all-or-nothing` option is indicated, but no explicit
value is passed.

We'll set a default option, which is retrieved in case the option
has not been provided, allowing us to override the value in case the
option is different than the default value.