Deprecate using `--all-or-nothing` with values
During the iteration of the PR https://github.com/doctrine/migrations/pull/1296 an idea to deprecate passing values to this option was suggested, given how complicated was to handle using that option without any value, which led to a bug and inconsistency with the documentation.

As specified in the deprecation message, using this option from 4.0.x onwards in the CLI will be treated as `true`, and the option itself won't accept values.