build: Fix cleanup of the backup composer lock file
The trap function was registered as the last bash instruction and as a
result was not executed if anything failed before, e.g. the composer
update command or building the PHAR itself.

In the event the build PHAR script fails, the temporary backup
composer.lock file may be created and not removed.