GH-1379: Improve Deprecation thrown check + logic
Unfortunately as part of the previous improvement to determine whether a deprecation is thrown, validating that the deprecation was not thrown ended up not being added explicitly.

In addition, adds the explicit expectation of not throwing the deprecation whenever the `--all-or-nothing` is not indicated, which was the root cause originating the issue.