GH-1443: fix `--all-or-nothing` default behavior
As result of some past changes to improve the way the option was inferred, several attempts to build a mechanism to handle the absence of value of this option were built via the following commits:

4da00c540672d687b8efa7b5f0c2b50107845379
533595134247c0c9f2e16c9a17fd1776a30303a4
5038099da818718703586804cc5bf890aec9086f
799f16d458a93aaadaec29012d8b8b425131fa1d
f726a5fdfeb9f61db594b436b62dbeeb5e925308

This commit leverages all the iterative improvements introduced along the way, and setting the correct value when the option is correctly specified.

This should bring the behavior inline with what is currently documented, while additionally adding a documentation deprecation informing users that passing a value to that option won't be allowed in 4.x

As an extra addition, this commit also introduces the negated option `--no-all-or-nothing`. The intent is to leverage the native Symfony >= 5.3 [Negatable command options](https://symfony.com/blog/new-in-symfony-5-3-negatable-command-options) in the future once the optional value from `--all-or-nothing` is completely removed.